Move related targets to separate directory
